+++ desafio-14/carvalhais/rust/Cargo.toml
@@ -1,8 +1,15 @@
 [package]
-name = "mathexpr"
+name = "numeric_expressions"
 version = "0.1.0"
 edition = "2021"
+authors = ["André Carvalhais"]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "parser"
+path = "src/parser.rs"
+
+[[bin]]
+name = "solver"
+path = "src/main.rs"
 
 [dependencies]
