+++ desafio-14/carvalhais/rust/src/parser.rs
@@ -1,0 +1,11 @@
+mod computer;
+mod scanner;
+mod shuntyard;
+mod types;
+
+pub mod prelude {
+    pub use super::computer::compute;
+    pub use super::scanner::scan;
+    pub use super::shuntyard::shunt;
+    pub use super::types::{Direction, Procedure, Token};
+}
